+++ src/app/main/declarations/dsp-dataDog-config.ts
@@ -1,0 +1,7 @@
+export class DspDataDogConfig {
+    constructor(public dataDogLogging: boolean,
+        public dataDogApplicationId: string,
+        public dataDogClientToken: string,
+        public dataDogSite: string,
+        public dataDogService: string) { }
+}