+++ src/app/workspace/resource/resource-instance-form/select-ontology/select-ontology.component.spec.ts
@@ -88,7 +88,7 @@ describe('SelectOntologyComponent', () => {
 
     it('should initialise the ontologies metadata', () => {
         expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies).toBeDefined();
-        expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies.length).toEqual(14);
+        expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies.length).toEqual(15);
     });
 
     it('should init the MatSelect and MatOptions correctly', async () => {
@@ -103,7 +103,7 @@ describe('SelectOntologyComponent', () => {
 
         const options = await select.getOptions();
 
-        expect(options.length).toEqual(14);
+        expect(options.length).toEqual(15);
 
         const option1 = await options[0].getText();
 
@@ -111,7 +111,7 @@ describe('SelectOntologyComponent', () => {
 
         const option2 = await options[1].getText();
 
-        expect(option2).toEqual('A minimal ontology');
+        expect(option2).toEqual('The free test ontology');
 
     });