+++ src/app/workspace/search/advanced-search/search-select-ontology/search-select-ontology.component.spec.ts
@@ -86,7 +86,7 @@ describe('SearchSelectOntologyComponent', () => {
     it('should initialise the ontologies\' metadata', () => {
 
         expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies).toBeDefined();
-        expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies.length).toEqual(14);
+        expect(testHostComponent.selectOntology.ontologiesMetadata.ontologies.length).toEqual(15);
 
     });
 
@@ -96,13 +96,13 @@ describe('SearchSelectOntologyComponent', () => {
         const initVal = await select.getValueText();
 
         // placeholder
-        expect(initVal).toEqual('Select an Ontology');
+        expect(initVal).toEqual('Select a data model');
 
         await select.open();
 
         const options = await select.getOptions();
 
-        expect(options.length).toEqual(14);
+        expect(options.length).toEqual(15);
 
         const option1 = await options[0].getText();
 
@@ -110,7 +110,7 @@ describe('SearchSelectOntologyComponent', () => {
 
         const option2 = await options[1].getText();
 
-        expect(option2).toEqual('A minimal ontology');
+        expect(option2).toEqual('The free test ontology');
 
     });